Creating permission steps incrementally (#78)
* initial incremental stepper variant.

* track collection structure.

* Minor tweaks with ForLoopDesugaring & branch joins

* Minor refactor and table entry filters.

Fixed table entry filtering scheme that was causing some of the loop
examples to fail. Table entries are now first grouped by line number
which is also how they are collapsed, this seems to have broken a
different example, and the issue with closures still remains. Needed
still is documentation and addressing comments.

* Fix loop spans and ifs without and else.

Removed old algorithm alongside other useless utilities like the
DFSFinder that was unstable. No "join" segments are inserted for
branches which I will need to revisit.

* Add weird_exprs tests, updated loops.

* Updated tests and documentation.

* Update test doc for SegmentedMir.